# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Blaze Sports Intelligence API
  description: |
    Production-grade sports analytics platform API providing real-time data,
    analytics, recruiting information, and vision AI capabilities.
  version: 1.0.0
  contact:
    name: Blaze Sports Intelligence
    url: https://blazesportsintel.com
  license:
    name: Proprietary

servers:
  - url: https://api.blazesportsintel.com/v1
    description: Production server
  - url: https://staging-api.blazesportsintel.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server

tags:
  - name: Authentication
    description: User authentication and authorization
  - name: Players
    description: Player information and statistics
  - name: Teams
    description: Team information and rosters
  - name: Games
    description: Game schedules, scores, and events
  - name: Leagues
    description: League information and standings
  - name: Recruiting
    description: Recruiting profiles and rankings
  - name: Analytics
    description: Advanced analytics and computed features
  - name: Vision AI
    description: Biomechanical analysis and character assessment

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      required:
        - error
        - message
        - statusCode
        - timestamp
        - path
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        requestId:
          type: string
        details:
          type: object

    Player:
      type: object
      required:
        - id
        - firstName
        - lastName
        - fullName
      properties:
        id:
          type: string
        teamId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        jerseyNumber:
          type: integer
        position:
          type: string
        height:
          type: number
          description: Height in inches
        weight:
          type: number
          description: Weight in pounds
        birthDate:
          type: string
          format: date
        hometown:
          type: string
        highSchool:
          type: string
        college:
          type: string
        profileImageUrl:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Team:
      type: object
      required:
        - id
        - leagueId
        - name
      properties:
        id:
          type: string
        leagueId:
          type: string
        name:
          type: string
        shortName:
          type: string
        city:
          type: string
        state:
          type: string
        mascot:
          type: string
        logoUrl:
          type: string
        isActive:
          type: boolean

    Game:
      type: object
      required:
        - id
        - leagueId
        - homeTeamId
        - awayTeamId
        - scheduledAt
        - status
      properties:
        id:
          type: string
        leagueId:
          type: string
        homeTeamId:
          type: string
        awayTeamId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [SCHEDULED, LIVE, HALFTIME, FINAL, POSTPONED, CANCELLED]
        homeScore:
          type: integer
        awayScore:
          type: integer

    PaginationMeta:
      type: object
      required:
        - total
        - limit
        - offset
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

paths:
  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

  /players:
    get:
      summary: Get all players
      tags: [Players]
      parameters:
        - name: teamId
          in: query
          schema:
            type: string
        - name: position
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Player'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      summary: Create new player
      tags: [Players]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Player'
      responses:
        '201':
          description: Player created
        '401':
          $ref: '#/components/responses/Unauthorized'

  /players/{id}:
    get:
      summary: Get player by ID
      tags: [Players]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Player details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Player'
        '404':
          $ref: '#/components/responses/NotFound'

  /games:
    get:
      summary: Get all games
      tags: [Games]
      parameters:
        - name: leagueId
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [SCHEDULED, LIVE, FINAL]
      responses:
        '200':
          description: List of games

  /games/live:
    get:
      summary: Get live games
      tags: [Games]
      responses:
        '200':
          description: Currently live games

  /analytics/player/{playerId}:
    get:
      summary: Get player analytics
      tags: [Analytics]
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Player analytics data
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/Unauthorized'